refactor: mv Waveform to dtype
