+++ src/iokit/dtype/waveform.py
